+++ py/sdrl/subcmd/student.py
@@ -62,6 +62,7 @@ def execute(pargs: argparse.Namespace):
     if pargs.op:
         OP_CMDS[pargs.op](context)  # execute one command via lookup table, with duck-typed arg
     else:
+        sdrl.report.print_volume_report(context.course, author_mode=False)
         run_command_loop(context, menu=MENU, helptext=MENU_HELP, cmds=MENU_CMDS)